Migrate -listen and -server settings from QSettings to settings.json
This commit also removes the `node.args = nullptr` assignment in the Shutdown()
function. Clearing node.args there never made sense because it made the
Shutdown() function not idempotent, making it fragile and causing issues like
https://github.com/bitcoin/bitcoin/issues/23186. Removing that assignment is
necessary in this commit to avoid a segfault in OptionTests during the new
node().initParameterInteraction() call added here. AppTests calls Shutdown()
which sets node.args to null, and OptionTests runs after AppTests and needs
node.args not to be null.