rpc: faster scanblocks by not traversing the whole chain block by block
The current flow walks-through every block in the active chain until
hits the chain tip or processes 10k blocks, then calls the
`lookupFilterRange` function to obtain all the filters from that
particular range.

Which isn't the best as it traverses the whole chain only to obtain
the heights range to lookup the block filters.

As `scanblocks` only lookup block filters in the active chain, we can
directly calculate the lookup range heights, by using the chain tip,
without requiring to traverse the chain block by block.