test: Coin Selection, duplicated preset inputs selection
This exercises the bug inside CoinsResult::Erase that
ends up on (1) a wallet crash or (2) a created and
broadcasted tx that pays an insane fee amount.

This is covered by making the wallet selects the preset
inputs twice during the coin selection process (which
is a dangerous bug on v24.0).

Making the wallet think that the selection process result covers
the entire tx target when it does not. It's actually creating
a tx that sends more coins than what inputs are covering for.

Which, combined with the subtract fee from outputs option,
makes the wallet generate and broadcast a tx that pays
up to the sum of all the preset inputs, minus one, amounts in
fee :(.

----------

Note:
To make review simpler, have coded the same test cases twice.
First as a unit test and secondly, inside a functional test
too.