wallet: regression fix, getAvailableBalance skips selected coins
Only a GUI issue, for the "use available balance" button when
the user manually select inputs on the send screen.

The previous behavior for getAvailableBalance when the coin
control has selected inputs was to return the sum of them. Instead,
we are currently returning the wallet's available total balance minus
the selected coins total amount.

Reason:
We missed to update the GetAvailableBalance function to include
the coin control selected coins on #25685.

Context:
Since #25685 we skip the selected coins inside `AvailableCoins`,
the reason is that there is no need to walk through the entire
wallet's txes map just to get coins that could have gotten by
just doing a simple mapWallet.find).